requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
aspose-words==25.4.0
asttokens==3.0.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
dotenv==0.9.9
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
frozenlist==1.5.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.69.1
greenlet==3.2.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy_extensions==1.1.0
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.93.0
orjson==3.10.18
packaging==24.2
parso==0.8.4
pdf2image==1.17.0
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-settings==2.10.1
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
python-pptx==1.0.2
# pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
rarfile==4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.41.3
tenacity==9.1.2
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
XlsxWriter==3.2.2
yarl==1.18.3
zstandard==0.23.0
